+++ crates/miden-lib/src/transaction/outputs.rs
@@ -18,8 +18,11 @@ pub const OUTPUT_NOTES_COMMITMENT_WORD_IDX: usize = 0;
 /// The index of the word at which the account update commitment is stored on the output stack.
 pub const ACCOUNT_UPDATE_COMMITMENT_WORD_IDX: usize = 1;
 
+/// The index of the word at which the fee asset is stored on the output stack.
+pub const FEE_ASSET_WORD_IDX: usize = 2;
+
 /// The index of the item at which the expiration block height is stored on the output stack.
-pub const EXPIRATION_BLOCK_ELEMENT_IDX: usize = 8;
+pub const EXPIRATION_BLOCK_ELEMENT_IDX: usize = 12;
 
 // ACCOUNT HEADER EXTRACTOR
 // ================================================================================================