+++ crates/miden-objects/src/transaction/outputs.rs
@@ -4,6 +4,7 @@ use core::fmt::Debug;
 use crate::{
     Felt, Hasher, MAX_OUTPUT_NOTES_PER_TX, TransactionOutputError, Word,
     account::AccountHeader,
+    asset::FungibleAsset,
     block::BlockNumber,
     note::{
         Note, NoteAssets, NoteHeader, NoteId, NoteMetadata, NoteRecipient, PartialNote,
@@ -24,6 +25,8 @@ pub struct TransactionOutputs {
     pub account_delta_commitment: Word,
     /// Set of output notes created by the transaction.
     pub output_notes: OutputNotes,
+    /// The fee of the transaction.
+    pub fee: FungibleAsset,
     /// Defines up to which block the transaction is considered valid.
     pub expiration_block_num: BlockNumber,
 }
@@ -33,6 +36,7 @@ impl Serializable for TransactionOutputs {
         self.account.write_into(target);
         self.account_delta_commitment.write_into(target);
         self.output_notes.write_into(target);
+        self.fee.write_into(target);
         self.expiration_block_num.write_into(target);
     }
 }
@@ -42,12 +46,14 @@ impl Deserializable for TransactionOutputs {
         let account = AccountHeader::read_from(source)?;
         let account_delta_commitment = Word::read_from(source)?;
         let output_notes = OutputNotes::read_from(source)?;
+        let fee = FungibleAsset::read_from(source)?;
         let expiration_block_num = BlockNumber::read_from(source)?;
 
         Ok(Self {
             account,
             account_delta_commitment,
             output_notes,
+            fee,
             expiration_block_num,
         })
     }